C and C++ programs to use calloc(), malloc() and free()
First of all, the prototypes of rmalloc(), calloc(), and free() are present in stdlib.h. Therefore, this file should be included as a header file.

Next, the syntax of malloc() is
	void *ptr = malloc(N * sizeof(type));
and the syntax of calloc is
	void *ptr = calloc(N, sizeof(type));

Now, the difference between malloc() and calloc() is:
	1. The syntax.
	2. The block freed up by malloc() for our use is not initialized with zero, but the block of memory allocated to us by calloc() is.
However, in my experiments on Ubuntu Linux, both malloc() and calloc() return block of memory whose bytes have been initialized to zero. This is a protection feature by the operating system's implementation of malloc() and calloc().

Next, to free up the block of memory and give it back to the operating system we make use of the free() function. The syntax of free() is
	free(void *)
So, you might need to explicitly convert the pointer to a void type before passing it as an argument to free().

Furthermore, invoking free() more than once with the same pointer can result in cache error.

Note that when a block of memory is allocated to the user by malloc() or calloc(), a header is attached to these blocks which specify their size along with other relevant information. So, when we pass a pointer to a block as an argument to free(), it looks for the size of the block which is provided in the header and frees up memory accordingly.

Andddd, say you have pointer int *ptr;
ptr + i will point to the block of memory with the address <ptr> + <i * sizeof(type of ptr)>. Therefore, you can add and subtract numbers to a pointer to point to various blocks having a size equal to the sizeof the type of variable to which the pointer points to. You need not worry about overlapping blocks of memory. This is a tongue twister, but it is easy to comprehend when you will look at the code.