Used lists to differentiate between iterables and iterators
A list can be iterated which makes it an iterable. But it itself cannot iterate through its elements, for which we need an iterator. To get an iterator for a list (or any other iterable object), pass that object to the iter() built-in function. Then use next() to iterate through the iterable object.