Code showing the way decorator is called
Decorators above a function are called every time the function is called. There is no exception to this rule. Even if the function calls itself recursively to perform some computation, the decorator will be called and then the function will be called from the decorator every time a recursive call is made.